refactor(hooks/useCardNumber): 카드 포맷팅 로직 분리
